pyproject.toml
[project]
name = "docx-compare"
version = "1.0.0"
description = "Sistema completo para comparação de documentos DOCX com integração Directus"
authors = [
    {name = "Devix Tecnologia", email = "contato@devix.co"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "directus-sdk-py>=1.1.0",
]

[project.scripts]
docx-compare = "src.docx_compare.processors.processador_automatico:main"
docx-diff = "src.docx_compare.core.docx_diff_viewer:main"

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused-function-argument
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "Q", # flake8-quotes
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "B904", # Within an `except` clause, raise exceptions with `raise ... from err`
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG001", "S101"]
"docx_diff_viewer.py" = ["W291"]  # Ignorar trailing whitespace no CSS embutido

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
]
